Update print statements and logger messages
Move ImageCallBack to lib

Insert doc string in sims

Move DonkeySimMsgHandler into dedicated class
