+++ perturbationdrive/__init__.py
@@ -63,7 +63,13 @@ from perturbationdrive.perturbationfuncs import (
 )
 
 from .utils.data_utils import CircularBuffer
-from .utils.logger import CSVLogHandler
+from .utils.logger import (
+    CSVLogHandler,
+    GlobalLog,
+    LOGGING_LEVEL,
+    ScenarioOutcomeWriter,
+    OfflineScenarioOutcomeWriter,
+)
 from .utils.utilFuncs import download_file, calculate_velocities
 from .SaliencyMap.saliencymap import (
     getActivationMap,
@@ -91,3 +97,4 @@ from .RoadGenerator.RandomRoadGenerator import RandomRoadGenerator
 from .RoadGenerator.CustomRoadGenerator import CustomRoadGenerator
 from .Simulator.Simulator import PerturbationSimulator
 from .Simulator.Scenario import Scenario, ScenarioOutcome, OfflineScenarioOutcome
+from .Simulator.image_callback import ImageCallBack
